Fix spacing, icon, color in snippet card body elements
Include an async toggle to switch a faved/unfaved snippet without relying on an asynchronous call and/or a fetch that would reload the whole page. Source and Edit buttons are modified to hide text in tight screens. Additional padding configs should be revised to make room for long tag names in that row